+++ docs/source/examples/plot_ukf_demo.py
@@ -97,7 +97,7 @@ sat1_agent = TargetAgent(
 # ---------------------------------------------
 #
 from resonaate.agents.estimate_agent import EstimateAgent
-from resonaate.estimation.sequential.unscented_kalman_filter import UnscentedKalmanFilter
+from resonaate.estimation.kalman.unscented_kalman_filter import UnscentedKalmanFilter
 from resonaate.physics.noise import continuousWhiteNoise, initialEstimateNoise
 from resonaate.scenario.config.estimation_config import InitialOrbitDeterminationConfig