+++ tests/sensors/test_field_of_view.py
@@ -13,8 +13,8 @@ from numpy import array, zeros
 from resonaate.agents.estimate_agent import EstimateAgent
 from resonaate.agents.sensing_agent import SensingAgent
 from resonaate.dynamics.two_body import TwoBody
+from resonaate.estimation.kalman.unscented_kalman_filter import UnscentedKalmanFilter
 from resonaate.estimation.maneuver_detection import StandardNis
-from resonaate.estimation.sequential.unscented_kalman_filter import UnscentedKalmanFilter
 from resonaate.physics.time.stardate import ScenarioTime
 from resonaate.physics.transforms.methods import getSlantRangeVector
 from resonaate.scenario.clock import ScenarioClock